moved firebase connection initialization to FirebaseConfig class
